checkpoint

+++ src/cache.py
@@ -10,40 +10,34 @@ class DOIReference:
         if isinstance(doi_input, DOIReference):
             # Clone the DOIReference object
             self._prefix = doi_input._prefix
-            self._postfix = doi_input._postfix
+            self._suffix = doi_input._suffix
         else:
             self.doi_input = doi_input
-            self._prefix, self._postfix = self._parse_doi(doi_input)
+            self._prefix, self._suffix = self._parse_doi(doi_input)
 
     def _parse_doi(self, doi_input):
-        """
-        Parse the input to extract prefix and postfix.
-        """
         if '/' in doi_input:
             if doi_input.startswith('https://doi.org/'):
                 doi_input = doi_input[len('https://doi.org/'):]
             parts = doi_input.split('/')
-            if len(parts) == 2:
-                return parts[0], parts[1]
-            else:
-                raise ValueError("Invalid DOI format")
+            return (parts[0], '/'.join(parts[1:]))
         else:
-            raise ValueError("Invalid DOI input, must contain '/' to separate prefix and postfix")
+            raise ValueError("Invalid DOI input, must contain '/' to separate prefix and suffix")
 
     @property
     def url(self):
         """
-        Construct the full DOI URL from prefix and postfix.
+        Construct the full DOI URL from prefix and suffix.
         """
-        return f"https://doi.org/{self._prefix}/{self._postfix}"
+        return f"https://doi.org/{self._prefix}/{self._suffix}"
 
     @property
     def prefix(self):
         return self._prefix
 
     @property
-    def postfix(self):
-        return self._postfix
+    def suffix(self):
+        return self._suffix
 
 
 def cache_result(filename_template):
@@ -52,7 +46,7 @@ def cache_result(filename_template):
         def wrapper(self, *args, **kwargs):
             filename = filename_template.format(
                 prefix=self.doi_reference.prefix,
-                postfix=self.doi_reference.postfix
+                suffix=self.doi_reference.suffix
             )
             cache_path = os.path.join(self.cache_path, filename)
             if not os.path.exists(cache_path):
@@ -73,8 +67,13 @@ class DOI:
         if not os.path.exists(self.cache_path):
             os.makedirs(self.cache_path)
 
-    @cache_result("{prefix}_{postfix}.pdf")
-    def get_pdf(self, pdf_path):
+    def format_filename(self, stem):
+        suffix = self.doi_reference.suffix.split('/')
+        suffix = '_'.join(suffix)
+        return f"{self.doi_reference.prefix}_{suffix}_{stem}"
+
+    def get_pdf(self):
+        return
         """
         Returns the path to the cached PDF, downloading it if necessary.
         """
@@ -84,19 +83,24 @@ class DOI:
             f.write("This is a dummy PDF content for DOI: " + self.doi_reference.url)
         #print(f"Downloaded PDF for DOI {self.doi_reference.url} to {pdf_path}")
 
-    @cache_result("{prefix}_{postfix}_crossref.json")
-    def get_crossref_data(self, json_path):
-        """
-        Returns the path to the cached Crossref data, downloading it if necessary.
-        """
-        work = xref.Works()
-        info = work.doi(self.doi_reference.url)
-        with open(json_path, 'w') as fh:
-            json.dump(info, fh, indent=2)
-        #print(f"Downloaded Crossref data for DOI {self.doi_reference.url} to {json_path}")
+    def load_xref_data(self):
+        filename = self.format_filename('xref.json')
+        filepath = os.path.join(self.cache_path, filename)
+        if os.path.exists(filepath):
+            with open(filepath, 'r') as fh:
+                info = json.load(fh)
+        else:
+            work = xref.Works()
+            info = work.doi(self.doi_reference.url)
+            with open(filepath, 'w') as fh:
+                json.dump(info, fh, indent=2)
+        return info
 
-    # Additional cacheable methods can be added here, e.g., get_ascii_text(), get_metadata(), etc.
-
+    @property
+    def xref(self):
+        if not hasattr(self, '_xref'):
+            self._xref = self.load_xref_data()
+        return self._xref
 
 class DOIFactory:
     def __init__(self, base_path='./doi-cache'):
@@ -113,23 +117,23 @@ class DOIFactory:
         return DOI(cache_path=cache_dir, doi_reference=doi_reference)
 
     def _get_cache_directory(self, doi_reference):
-        """
-        Construct the cache directory based on the DOI.
-        """
-        return os.path.join(self.base_path, doi_reference.prefix, doi_reference.postfix)
-
+        suffix = doi_reference.suffix.split('/')
+        suffix = '_'.join(suffix)
+        return os.path.join(self.base_path, doi_reference.prefix, suffix)
 
 # Example usage
 if __name__ == "__main__":
     from doi_model import ResearchDocument
     factory = DOIFactory()
     #doi_obj = factory.create_doi("https://doi.org/10.1101/2023.12.12.571258")
-    doi_obj = factory.create_doi("https://doi.org/10.1101/460345")
+    #doi_obj = factory.create_doi("https://doi.org/10.1101/460345")
+    #doi_obj = factory.create_doi("https://doi.org/10.1101/2022.04.21.488948")
+    doi_obj = factory.create_doi("https://doi.org/10.1093/bioinformatics/btac729")
     pdf_path = doi_obj.get_pdf()
     #print(f"PDF cached at: {pdf_path}")
-    crossref_path = doi_obj.get_crossref_data()
+    import pprint
+    pprint.pprint(doi_obj.xref)
     #print(f"Crossref data cached at: {crossref_path}")
-    with open(crossref_path) as fh:
-        obj = json.load(fh)
-    doc = ResearchDocument(**obj)
-    print(json.dumps(json.loads(doc.json())))
+    doc = ResearchDocument(**doi_obj.xref)
+    #print(doc.json())
+    #print(json.dumps(json.loads(doc.json())))
